now prints as ints instead of double

+++ trainGameJava/four_Ten.java
@@ -11,24 +11,24 @@ public class four_Ten {
             for(int j = 0; j < 2; j++){          
                  
                 if(abc.results[i] + d.results[j] == 10){
-                    buffer = String.format("%s + %f\n", abc.strings.get(i), d.results[j]);
+                    buffer = String.format("%s + %d\n", abc.strings.get(i), (int)d.results[j]);
                     s.addString(buffer);
                 }
         
                 if(abc.results[i] - d.results[j] == 10){
-                    buffer = String.format("%s - %f\n", abc.strings.get(i), d.results[j]);
+                    buffer = String.format("%s - %d\n", abc.strings.get(i), (int)d.results[j]);
                     s.addString(buffer);     
                 }
                 if(abc.results[i] * d.results[j] == 10){
-                    buffer = String.format("%s * %f\n", abc.strings.get(i), d.results[j]);
+                    buffer = String.format("%s * %d\n", abc.strings.get(i), (int)d.results[j]);
                     s.addString(buffer);
                 }
                 if(abc.results[i] / d.results[j] == 10){
-                    buffer = String.format("%s / %f\n", abc.strings.get(i), d.results[j]);
+                    buffer = String.format("%s / %d\n", abc.strings.get(i), (int)d.results[j]);
                     s.addString(buffer);
                 }
                 if(Math.pow((abc.results[i]),(d.results[j])) == 10){
-                    buffer = String.format("%s ^ %f\n", abc.strings.get(i), d.results[j]);
+                    buffer = String.format("%s ^ %d\n", abc.strings.get(i), (int)d.results[j]);
                     s.addString(buffer);
                 }
             }
